Fix an event parameter passing bug
In the previous version, this method captured a user defined event in the callback function and caused a bug when an event parameter was disposed when it was passed to the user defined onClick handler (read at https://reactjs.org/docs/events.html). Also added a feature to prevent a default event, in order to do it just return false from the custom onClick handler.